chore: re-organize directories
