+++ src/components/PokemonPhysical.jsx
