+++ src/components/PokemonStats.jsx
